+++ src/_app/providers/Theme/index.ts
@@ -1,0 +1,1 @@
+export { Theme } from './Theme'
