+++ src/entities/postsList/hook/useGetAllPosts.ts
@@ -2,6 +2,8 @@ import { useCallback, useMemo } from 'react'
 
 import { useGetPostsQuery } from '@/entities/postsList/api/getPosts.api.types'
 
+//https://dev.to/moruno21/graphql-infinite-scroll-4oan
+
 export const useGetAllPosts = () => {
   const { data, fetchMore, loading } = useGetPostsQuery({
     variables: {
@@ -21,7 +23,6 @@ export const useGetAllPosts = () => {
     if (posts.length === totalCount) {
       return
     }
-
     fetchMore({
       updateQuery: (previousResult, { fetchMoreResult }) => {
         if (!fetchMoreResult) {
