+++ src/entities/postsList/ui/PostDescription/index.ts
@@ -1,0 +1,1 @@
+export { PostDescription } from './PostDescription'
