+++ src/entities/postsList/ui/PostItem/index.ts
@@ -1,0 +1,1 @@
+export { PostItem } from './PostItem'
