+++ src/entities/postsList/ui/PostPhotos/index.ts
@@ -1,0 +1,1 @@
+export { PostPhotos } from './PostPhotos'
