+++ src/entities/postsList/ui/PostUserInfo/index.ts
@@ -1,0 +1,1 @@
+export { PostUserInfo } from './PostUserInfo'
