+++ src/entities/postsList/ui/Posts/Posts.tsx
@@ -1,12 +1,14 @@
 import InfiniteScroll from 'react-infinite-scroll-component'
 
 import { useGetAllPosts } from '@/entities/postsList/hook/useGetAllPosts'
-import { PostsStyled } from '@/entities/postsList/ui/Posts/Posts.styled'
-import { PostsList } from '@/entities/postsList/ui/PostsList/PostsList'
+import { PostsList } from '@/entities/postsList/ui/PostsList'
 import { PostsListSkeleton } from '@/entities/postsList/ui/PostsListSkeleton'
+import { IsEmpty } from '@/shared/components/IsEmpty'
 import { useTranslation } from '@/shared/hooks/useTranslation'
 import { ScrollToTop } from '@/shared/ui/ScrollToTop/ScrollToTop'
-import { InputText, Skeleton } from '@nazar-pryt/inctagram-ui-kit'
+import { InputText } from '@nazar-pryt/inctagram-ui-kit'
+
+import { PostsStyled } from './Posts.styled'
 
 export const Posts = () => {
   const { fetchMoreData, hasMore, isHavePosts, loading, posts } = useGetAllPosts()
@@ -38,5 +40,5 @@ export const Posts = () => {
     )
   }
 
-  return <div>{t.post.no_posts}</div>
+  return <IsEmpty text={t.post.no_posts} />
 }
