+++ src/entities/postsList/ui/PostsList/index.ts
@@ -1,0 +1,1 @@
+export { PostsList } from './PostsList'
